alsaseq: fix constant objects in Python 3
In Py3 everything is a long, so we can't get away with duplicating the
integer object layout any more. Properly subclass PyLong_Type instead,
with a bit of a hack to add an extra name field after the variable-sized
data section.

Also get rid of the enum type stuff, which seems to be unused cruft, and
the numprotocol stuff, because the base type can take care of that (and
we can't map arbitrary arithmetic results back to constants anyway).

Signed-off-by: Hector Martin <marcan@marcan.st>
Signed-off-by: Takashi Iwai <tiwai@suse.de>